Main

Documentation

main :: IO ()

The beginning of everything.

ui

Arguments

:: [String]
List with words from text

-> Int
Indicates how many words should be displayed

-> Tree (Char, Bool)
Wordlist as a Trie

-> Bool
Indicates whether to use matrices or not

-> IO ()

getCorrectedWords

Arguments

:: String
Typo

-> Tree (Char, Bool)
Wordlist as a Trie

-> Bool
indicates whether to use matrices or not

-> [String]
List with suggestions for typo

Gives a list of suggestions for a typo.

stringArray :: [String] -> [(Char, [(Char, Int)])]

Transforms all word list from a file to a two dimensional array.

getMatrixWithCosts :: [(Char, [(Char, Int)])] -> [(Char, [(Char, Int)])]

Change matrix with absolute frequency to matrix with costs

transformToCosts :: Int -> Int -> Int

Transforms a values for the frequency of yx and x to costs.

checkWord :: String -> Tree (Char, Bool) -> Bool

Checking whether a word is already in the trie .

insertAll :: [String] -> Tree (Char, Bool)

Insert a list of words into the trie.

insertT :: String -> Tree (Char, Bool) -> Tree (Char, Bool)

insert a word into the trie

lookupT :: Char -> [Tree (Char, Bool)] -> Maybe Int

Returns Just the index of a char in a root node or nothing.

updateL :: [Tree (Char, Bool)] -> Int -> (Tree (Char, Bool) -> Tree (Char, Bool)) -> [Tree (Char, Bool)]

Uses a function on the nth element in a list of tries.

stringToTree :: String -> Tree (Char, Bool)

Converts a word in to a list trie.

readOff :: Tree (Char, Int, Bool) -> [(String, Int)]

Returns all words in a trie. (with duplicates)

changeN :: Tree (Char, Bool) -> Tree (Char, Bool)

Sets the root node of a trie to true.

min :: Ord a => [(t, a)] -> (t, a)

Returns the minimal tuple of a list of tuples

takeWithoutDups :: Eq a => Int -> [a] -> [a]

Take the first n elements from a list without duplicates